requirements.txt
APScheduler==3.10.4
Flask==2.2.2
python-dotenv==1.0.0
Requests==2.31.0
